Force parsing of extracted order (#798)
Features:
- Forced serialization of returned response from extractor call.
- Sometimes it returns a `str` which we previously threw an error for
because that would indicate it failed to extract. But because we
introduced a lot of new constraints (e.g. first name, phone_number,
order_type, etc. must be provided by the user) sometimes the extractor
fails to extract a suitable `Order` object and returns a `str` with
semi-filled data. So instead, if we get either a `str` or `Order` we
accept it. For the case of a `str`, we manually parse it into a `Order`
object.

- Updated extractor prompts to better handle extracting takeout or
delivery.
- Updated chat history to only include assistant content not other
`extra` fields
- Added phone_number formatter upon `Order` model serialization.


## Checklist

### Before merging
- [x] Test: I have tested my changes locally and verified the changes
work as expected. [api](http://localhost:8000/docs),
[app](http://localhost:8501/)
### After merging
- [ ] Deploy: I have verified the deployment was successful. Deployment
status is updated in
[#devops-release-lat-pal-mono](https://proactiveailabgroup.slack.com/archives/C07H6H6MVV0).
- [ ] Test: I have verified that the changes work as expected on `lat`
environment.
[api](http://pal-mono-lat-api-lb-1443082111.us-west-1.elb.amazonaws.com/docs),
[app](http://pal-mono-lat-app-lb-1258791823.us-west-1.elb.amazonaws.com/)
- [ ] Monitor: I have monitored the `lat` environment stay healthy for
10 minutes after deployment. Health status is updated in
[#devops-health-lat](https://proactiveailabgroup.slack.com/archives/C07KAU9AUBC).