Fix db migration issues (#799)
- The alembic db migration always recreates existing tables which is not
fun. This PR attempts at fixing that problem by removing the specific
"public" schema in the Base class. It seems like when the "public"
schema is specified, alembic does something funky when it generates the
migration files. Feel free to pull this branch and apply the migration.
Then play around with modifying the model files to see that it (should)
only generate the actual diff in the migration file.
- Commented out the post_write_hooks since we don't have a script for
that so we don't see that error message at the end of autogenerate.

## Checklist

### Before merging
- [ ] Test: I have tested my changes locally and verified the changes
work as expected. [api](http://localhost:8000/docs),
[app](http://localhost:8501/)
### After merging
- [ ] Deploy: I have verified the deployment was successful. Deployment
status is updated in
[#devops-release-lat-pal-mono](https://proactiveailabgroup.slack.com/archives/C07H6H6MVV0).
- [ ] Test: I have verified that the changes work as expected on `lat`
environment.
[api](http://pal-mono-lat-api-lb-1443082111.us-west-1.elb.amazonaws.com/docs),
[app](http://pal-mono-lat-app-lb-1258791823.us-west-1.elb.amazonaws.com/)
- [ ] Monitor: I have monitored the `lat` environment stay healthy for
10 minutes after deployment. Health status is updated in
[#devops-health-lat](https://proactiveailabgroup.slack.com/archives/C07KAU9AUBC).

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Chores**
- Disabled an automated post-action configuration related to the `black`
formatter and removed related linting configurations.
- **New Features**
- Introduced a migration script to rename indexes in various database
tables for improved clarity.
- **Refactor**
- Streamlined the database model setup by removing the custom base class
and simplifying schema management.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Frankie Liu <frankie.liu@proactiveailab.com>