Remove dependency on legacy tools (#886)
Summary:
Deprecating the legacy tools. The first step is to decouple its
dependency.
The legacy tools exposes two functions: 
  generate_output_model and get_tools
Both are used by legacy agent, which is being deprecating next. 
For simplicity, we will not give tools to legacy agent.

Another dependency is in the app where BookingTools are directly used.
Replace it with the new BookingTool. This is also not needed since "app"
is dying now.


## Checklist

### Before merging
- [ ] Test: I have tested my changes locally and verified the changes
work as expected. [api](http://localhost:8000/docs),
[app](http://localhost:8501/)
### After merging
- [ ] Deploy: I have verified the deployment was successful. Deployment
status is updated in
[#devops-release-lat-pal-mono](https://proactiveailabgroup.slack.com/archives/C07H6H6MVV0).
- [ ] Test: I have verified that the changes work as expected on `lat`
environment.
[api](http://pal-mono-lat-api-lb-1443082111.us-west-1.elb.amazonaws.com/docs),
[app](http://pal-mono-lat-app-lb-1258791823.us-west-1.elb.amazonaws.com/)
- [ ] Monitor: I have monitored the `lat` environment stay healthy for
10 minutes after deployment. Health status is updated in
[#devops-health-lat](https://proactiveailabgroup.slack.com/archives/C07KAU9AUBC).

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Streamlined the response handling in the agent integration process for
consistent, non-stream output.
- Simplified the booking tool initialization on the development page for
a more efficient and straightforward experience.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->